moved firebase config to its own component and imported it into jobs
